small cleanup and restructure
